+++ src/base/b-list/interface.ts
@@ -86,3 +86,4 @@ export interface Item extends Super {
 }
 
 export type Items = Item[];
+