+++ src/traits/i-active-items/i-active-items.ts
@@ -166,6 +166,10 @@ export default abstract class iActiveItems extends iItems {
 
 	/** @see [[iActiveItems.prototype.setActive]] */
 	static setActive(ctx: TraitComponent, value: iActiveItems['Active'], unsetPrevious?: boolean): boolean {
+		if (!this.isActivatable(ctx, value)) {
+			return false;
+		}
+
 		const
 			activeStore = ctx.field.get('activeStore');
 
@@ -264,6 +268,10 @@ export default abstract class iActiveItems extends iItems {
 
 	/** @see [[iActiveItems.prototype.toggleActive]] */
 	static toggleActive(ctx: TraitComponent, value: iActiveItems['Active'], unsetPrevious?: boolean): iActiveItems['Active'] {
+		if (!this.isActivatable(ctx, value)) {
+			return false;
+		}
+
 		const
 			activeStore = ctx.field.get('activeStore');
 
@@ -300,6 +308,22 @@ export default abstract class iActiveItems extends iItems {
 		}
 
 		return ctx.active;
+	}
+
+	/** @see [[iActiveItems.prototype.getItemByValue]] */
+	static getItemByValue(ctx: TraitComponent, value: Item['value']): CanUndef<Item> {
+		return ctx.items?.find((item) => item.value === value);
+	}
+
+	/**
+	 * Checks if item can possibly be active by its value
+	 *
+	 * @param ctx
+	 * @param value
+	 */
+	protected static isActivatable(ctx: TraitComponent, value: Item['value']): boolean {
+		const item = ctx.getItemByValue(value);
+		return item?.activatable !== false;
 	}
 
 	/**
@@ -347,6 +371,15 @@ export default abstract class iActiveItems extends iItems {
 	 * @emits `immediateChange(active: unknown)`
 	 */
 	toggleActive(value: Item['value'], unsetPrevious?: boolean): iActiveItems['Active'] {
+		return Object.throw();
+	}
+
+	/**
+	 * Returns an item by the specified value
+	 *
+	 * @param value
+	 */
+	getItemByValue(value: Item['value']): CanUndef<Item> {
 		return Object.throw();
 	}
 }